Testing configured for backend and frontend
